# docs/swagger.yaml
definitions:
...
  controllers.accessmenuInput:
    properties:
      menuID:
        type: integer
      roleID:
        type: integer
    type: object
...
  controllers.dataMenu:
    properties:
      icon:
        type: string
      name:
        type: string
      title:
        type: string
      url:
        type: string
    type: object
  controllers.menuInput:
    properties:
      name:
        type: string
    type: object
...
  controllers.submenuInput:
    properties:
      icon:
        type: string
      menuID:
        type: integer
      title:
        type: string
      url:
        type: string
    type: object
  models.Accessmenu:
    properties:
      id:
        type: integer
      menuID:
        type: integer
      roleID:
        type: integer
    type: object
...
  models.Menu:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.Submenu:
    properties:
      icon:
        type: string
      id:
        type: integer
      menuID:
        type: integer
      title:
        type: string
      url:
        type: string
    type: object
info:
  contact: {}
paths:
  /accessmenu/:
    post:
      description: Creating a new Accessmenu.
      parameters:
      - description: the body to create a new Accessmenu
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.accessmenuInput'
      - default: Bearer
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Accessmenu'
      security:
      - ApiKeyAuth: []
      summary: Create New Accessmenu`.
      tags:
      - Accessmenu
  /accessmenu/{id}:
    delete:
      description: Delete a Accessmenu by id.
      parameters:
      - description: Accessmenu id
        in: path
        name: id
        required: true
        type: string
      - default: Bearer
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete one Accessmenu.
      tags:
      - Accessmenu
    patch:
      description: Update Accessmenu by id.
      parameters:
      - description: Accessmenu id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update age rating category
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.accessmenuInput'
      - default: Bearer
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Accessmenu'
      security:
      - ApiKeyAuth: []
      summary: Update Accessmenu.
      tags:
      - Accessmenu
...
  /menu/:
    post:
      description: Creating a new Menu.
      parameters:
      - description: the body to create new Menu
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.menuInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Menu'
      security:
      - BearerToken: []
      summary: Create New Menu`.
      tags:
      - Menu
  /menu/{id}:
    delete:
      description: Delete a Menu by id.
      parameters:
      - description: Menu id
        in: path
        name: id
        required: true
        type: string
      - default: Bearer
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete one Menu.
      tags:
      - Menu
    patch:
      description: Update Menu by id.
      parameters:
      - description: Menu id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update age rating category
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.menuInput'
      - default: Bearer
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Menu'
      security:
      - ApiKeyAuth: []
      summary: Update Menu.
      tags:
      - Menu
  /menu/role/{id}:
    get:
      description: Get an Menu by Role id.
      parameters:
      - description: Role id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.dataMenu'
      summary: Get Menu.
      tags:
      - Menu
...
  /submenu/:
    post:
      description: Creating a new Submenu.
      parameters:
      - description: the body to create a new Submenu
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.submenuInput'
      - default: Bearer
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Submenu'
      security:
      - ApiKeyAuth: []
      summary: Create New Submenu`.
      tags:
      - Submenu
  /submenu/{id}:
    delete:
      description: Delete a Submenu by id.
      parameters:
      - description: Submenu id
        in: path
        name: id
        required: true
        type: string
      - default: Bearer
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete one Submenu.
      tags:
      - Submenu
    get:
      description: Get an Submenu by id.
      parameters:
      - description: Submenu id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Submenu'
      summary: Get Submenu.
      tags:
      - Submenu
    patch:
      description: Update Submenu by id.
      parameters:
      - description: Submenu id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update age rating category
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.submenuInput'
      - default: Bearer
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Submenu'
      security:
      - ApiKeyAuth: []
      summary: Update Submenu.
      tags:
      - Submenu
swagger: "2.0"